Revert "bump revm-inspectors + type casting to u128"
This reverts commit 25aa23cffaadef1d047ce7b359b4d7ad5018704a.

+++ crates/anvil/core/src/eth/transaction/optimism.rs
@@ -178,8 +178,8 @@ impl Transaction for DepositTransactionRequest {
     }
 
     /// Get `gas_limit`.
-    fn gas_limit(&self) -> u64 {
-        self.gas_limit as u64
+    fn gas_limit(&self) -> u128 {
+        self.gas_limit
     }
 
     /// Get `gas_price`.
